FIX: Pages가 모든 페이지 컴포넌트를 포함하도록 변경
