+++ frontend/src/pages/About/AboutPage.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import About from "components/About";
+
+const AboutPage = () => {
+  return <About />;
+};
+
+export default AboutPage;
