--- frontend/src/pages/AboutPage.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import AboutContainer from "../containers/About/AboutContainer";
-
-const AboutPage = () => {
-  return (
-    <>
-      <AboutContainer />
-    </>
-  );
-};
-
-export default AboutPage;
