+++ frontend/src/pages/Auth/AuthPage.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import Auth from "components/Auth";
+
+const AuthPage = () => {
+  return <Auth />;
+};
+
+export default AuthPage;
