--- frontend/src/pages/AuthPage.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import AuthContainer from "../containers/Auth/AuthContainer";
-
-const AuthPage = () => {
-  return (
-    <>
-      <AuthContainer />
-    </>
-  );
-};
-
-export default AuthPage;
