+++ frontend/src/pages/Categories/CategoriesPage.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import Categories from "components/Categories";
+
+const CategoriesPage = () => {
+  return <Categories />;
+};
+
+export default CategoriesPage;
