--- frontend/src/pages/CategoriesPage.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import CategoriesContainer from "../containers/Categories/CategoriesContainer";
-
-const CategoriesPage = () => {
-  return (
-    <>
-      <CategoriesContainer />
-    </>
-  );
-};
-
-export default CategoriesPage;
