+++ frontend/src/pages/Handle/HandlePage.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import Handle from "components/Handle";
+
+const HandlePage = () => {
+  return <Handle />;
+};
+
+export default HandlePage;
