+++ frontend/src/pages/Handle/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./HandlePage";
