--- frontend/src/pages/HandlePage.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import HandleContainer from "../containers/Handle/HandleContainer";
-
-const HandlePage = () => {
-  return (
-    <>
-      <HandleContainer />
-    </>
-  );
-};
-
-export default HandlePage;
