+++ frontend/src/pages/Main/MainPage.tsx
@@ -1,0 +1,8 @@
+import Main from "components/Main";
+import React from "react";
+
+const MainPage = () => {
+  return <Main />;
+};
+
+export default MainPage;
