+++ frontend/src/pages/Main/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./MainPage";
