--- frontend/src/pages/MainPage.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import MainContainer from "../containers/Main/MainContainer";
-
-const MainPage = () => {
-  return (
-    <>
-      <MainContainer />
-    </>
-  );
-};
-
-export default MainPage;
