--- frontend/src/pages/MemberPage.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import MembersContainer from "../containers/Members/MembersContainer";
-
-const MemberPage = ({}) => {
-  return (
-    <>
-      <MembersContainer />
-    </>
-  );
-};
-
-export default MemberPage;
