+++ frontend/src/pages/Members/MembersPage.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import Members from "components/Members";
+
+const MembersPage = () => {
+  return <Members />;
+};
+
+export default MembersPage;
