+++ frontend/src/pages/Members/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./MembersPage";
