+++ frontend/src/pages/Post/PostPage.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import Post from "components/Post";
+
+const PostPage = () => {
+  return <Post />;
+};
+
+export default PostPage;
