+++ frontend/src/pages/Post/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./PostPage";
