--- frontend/src/pages/PostPage.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import PostContainer from "../containers/Post/PostContainer";
-
-const PostPage = () => {
-  return (
-    <>
-      <PostContainer />
-    </>
-  );
-};
-
-export default PostPage;
