+++ frontend/src/pages/Search/SearchPage.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import Search from "components/Search";
+
+const SearchPage = () => {
+  return <Search />;
+};
+
+export default SearchPage;
