--- frontend/src/pages/SearchPage.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import SearchContainer from "../containers/Search/SearchContainer";
-
-const SearchPage = () => {
-  return (
-    <>
-      <SearchContainer />
-    </>
-  );
-};
-
-export default SearchPage;
