+++ frontend/src/pages/Temp/TempPage.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import Temp from "components/Temp";
+
+const TempPage = () => {
+  return <Temp />;
+};
+
+export default TempPage;
