--- frontend/src/pages/TempPage.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import TempContainer from "../containers/Temp/TempContainer";
-
-const TempPage = () => {
-  return (
-    <>
-      <TempContainer />
-    </>
-  );
-};
-
-export default TempPage;
