+++ frontend/src/pages/index.ts
@@ -1,0 +1,9 @@
+export { default as About } from "./About";
+export { default as Auth } from "./Auth";
+export { default as Categories } from "./Categories";
+export { default as Handle } from "./Handle";
+export { default as Main } from "./Main";
+export { default as Members } from "./Members";
+export { default as Post } from "./Post";
+export { default as Search } from "./Search";
+export { default as Temp } from "./Temp";
